move ResultName, ResultColor and Color into separate file
